Add files via upload
This is a project developed by myself and others in our Machine Learning II course. The goal of this project was to use Computer Vision to identify characters on license plates to provide a cost efficient, less resource intensive solution to any of their possible daily operating bottlenecks. We developed two Neural Networks [Feed Forward, and Convolutional] of low complexities and cross examined their respective outputs. As we anticipated, the CNN outperformed the FFNN as it's generally used for image/video data related tasks in practice. We achieved an accuracy score of ~98% on the convolutional model. If you're interested in learning more or discussing our approaches, ideas, etc. feel free to reach out!